v1.1.0 - Collisions are no longer possible after start is called
Adjusted the way Fabricator works so that once start is called,
collisions with built in methods are no longer a possibility. This is
made possible by checking to see if Fabricator has been started when
each method is called. If start() has already been called, execution is
interrupted and handed off to the __getattr_started method. This looks
up the correct endpoint and then calls that endpoint.

Doing this required that all methods be converted to use *args and
**kwargs to receive their arguments -- to allow them to pass those
arguments on to the endpoint in the case where .start() has already been
called.